% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_week_prior.R
\name{eval_week_prior}
\alias{eval_week_prior}
\title{Evaluate predictions for a given week (for tuning)}
\usage{
eval_week_prior(preds, season, week)
}
\description{
Evaluate predictions for a given week (for tuning)
}
